rpc: move result package into response
`response` is to be populated soon.
