+++ pkg/rpc/response/result/asset_state.go
