+++ pkg/rpc/response/result/contract_state.go
