+++ pkg/rpc/response/result/peers_test.go
